24-11

+++ resources/js/Pages/Managements/Customers/Index.jsx
@@ -71,6 +71,9 @@ const Customer = ()=>{
             toast.warning('Por favor seleccione un registro para eliminar!');
             return;
         }
+        if (!confirm('¿Estás seguro de eliminar?')) {
+            return;
+        }
         axios.delete(route('customers.destroy',selectedRecord))
         .then(response => {
             toast.success(response.data.message);
